producer/requirements.txt
requests==2.31.0
kafka-python-ng==2.2.2
python-dotenv==1.0.0